+++ src/frontend/hooks/products/useProductsQuery.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { getProducts } from '../../services/getProducts'
+import { getProducts } from '../../services/products/getProducts'
 
 const useProductsQuery = (queryId: string) => {
   const query = useQuery([queryId], () => getProducts().then((res) => res.data))
